# docs/swagger.yaml
definitions:
  internal.message:
    properties:
      message:
        type: string
    type: object
  internal.todo:
    properties:
      id:
        type: string
      task:
        type: string
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
paths:
  /todo:
    get:
      operationId: get-all-todos
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal.todo'
      summary: get all items in the todo list
    post:
      operationId: create-todo
      parameters:
      - description: todo data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/internal.todo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal.todo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal.message'
      summary: add a new item to the todo list
  /todo/{id}:
    delete:
      operationId: delete-todo-by-id
      parameters:
      - description: todo ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal.todo'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal.message'
      summary: delete a todo item by ID
    get:
      operationId: get-todo-by-id
      parameters:
      - description: todo ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal.todo'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal.message'
      summary: get a todo item by ID
swagger: "2.0"